requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
networkx==3.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
pymongo==4.3.3
python-dotenv==1.0.0
pyvis==0.3.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
